introduce new energy model utils
